+++ linbox/tests/test-rational-solver.C
@@ -106,7 +106,7 @@ bool testRandomSolve (const Ring& R,
 		RSolver rsolver;
 
 		//std::vector<std::pair<typename Ring::Element, typename Ring::Element> > answer(n);
-		std::vector<typename Ring::Element> num(n);
+		std::vector<typename Ring::Element> num((size_t)n);
 		typename Ring::Element den;
 
 		SolverReturnStatus solveResult = rsolver.solve(num, den, D, b, 30); //often 5 primes are not enough
